Manipulator Work
May or may not be in some relative form of completion, further judgement
and discretion is recommended, if not required, for the personal safety
and sanity of all viewers and robots involved.  Shockwave is personally
not responsible for any damages, or loss of property as a result of the
use and/or abuse of the following code.  Any lawsuits made will be
immediately terminated and civil action may be taken upon those who
violate the following contract in any way.
Thank you for your participation in this survey
Neil DeGrasse Tyson