Dashboard work
You gave me a list of things to put into the dashboard, but some values
required the creation of some new functions. I was planning on having
all of the dashboard stuff in one function that could be called
periodically, to make the main class easier to look at.
Cheers,
Nic
PS: I hope this didn't break anything. It may be unfinished...